# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ROS OCP Backend API",
    "description": "GO Echo API for Resource Optimization OCP Backend Service.\n\n**Plugin system:** Endpoints are served by plugins that can be individually enabled or disabled via `ROS_ENABLED_PLUGINS` / `ROS_DISABLED_PLUGINS` environment variables. Disabled plugin routes return `404 Not Found`. Available plugins: `container`, `namespace`, `gpu`, `node`, `pvc`, `snapshot`, `kruize`. Default: all native plugins enabled, kruize disabled. See docs/architecture/plugin-architecture.md for details.\n\n**Term configuration:** Recommendation terms (short/medium/long) control the lookback window for each plugin. Terms can be customized per-tenant via PUT /settings/terms. Administrators can lock terms via environment variables (`ROS_TERMS_<PLUGIN>_<TERM>_WINDOW_DAYS`, `ROS_TERMS_<PLUGIN>_<TERM>_MIN_DATA_DAYS`, `ROS_TERMS_<PLUGIN>_<TERM>_DECAY_HALFLIFE_HOURS`). Locked terms cannot be modified by tenants (PUT returns 422). Maximum window_days varies by plugin: 90 for container/namespace/node/gpu, 365 for pvc.",
    "version": "v1",
    "license": {
      "name": "GPLv3",
      "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
    }
  },
  "servers": [
    {
      "url": "/api/cost-management/v1",
      "description": "Default API prefix"
    }
  ],
  "paths": {
...
    "/recommendations/openshift/settings/terms": {
      "get": {
        "tags": ["Settings"],
        "summary": "Get recommendation term settings",
        "description": "Returns the current recommendation term configuration for the caller's org and recommendation type. Resolves precedence: admin env var > tenant DB override > plugin default. Terms marked `locked: true` cannot be modified via the API.",
        "operationId": "getTermSettings",
        "parameters": [
          {
            "name": "recommendation_type",
            "in": "query",
            "required": true,
            "description": "The recommendation plugin to get terms for. Use GET /settings/capabilities to list valid values.",
            "schema": {
              "type": "string",
              "enum": ["container", "namespace", "node", "gpu", "pvc"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Term settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TermSettingsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid recommendation_type"
          }
        }
      },
      "put": {
        "tags": ["Settings"],
        "summary": "Update recommendation term settings",
        "description": "Replace the recommendation term configuration for the caller's org and recommendation type. Maximum window_days varies by plugin (90 for container/namespace/node/gpu, 365 for pvc). If min_data_days is omitted, it is auto-computed as half of window_days. Terms locked by administrator environment variables cannot be modified (returns 422).",
        "operationId": "putTermSettings",
        "parameters": [
          {
            "name": "recommendation_type",
            "in": "query",
            "required": true,
            "description": "The recommendation plugin to update terms for.",
            "schema": {
              "type": "string",
              "enum": ["container", "namespace", "node", "gpu", "pvc"]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TermSettingsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated term settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TermSettingsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error (invalid name, window_days out of range, min_data_days > window_days, etc.)"
          },
          "422": {
            "description": "One or more terms are locked by administrator environment variables and cannot be modified",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "error" },
                    "message": { "type": "string" },
                    "locked_terms": {
                      "type": "array",
                      "items": { "type": "string" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Settings"],
        "summary": "Reset recommendation term settings to defaults",
        "description": "Removes any custom term configuration for the caller's org and recommendation type, reverting to plugin defaults. If terms are admin-locked, the DELETE still succeeds (removes only the DB override; locked values remain active).",
        "operationId": "deleteTermSettings",
        "parameters": [
          {
            "name": "recommendation_type",
            "in": "query",
            "required": true,
            "description": "The recommendation plugin to reset terms for.",
            "schema": {
              "type": "string",
              "enum": ["container", "namespace", "node", "gpu", "pvc"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Default term settings (after reset)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TermSettingsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid recommendation_type"
          }
        }
      }
    },
...
    "/recommendations/openshift/gpu": {
      "get": {
        "tags": ["GPU Recommendations"],
        "summary": "GPU recommendations summary",
        "description": "Lightweight counts for MIG vs time-slicing listings, distinct GPU triples ever analyzed, and clusters with GPU digest data.",
        "operationId": "getGpuRecommendationsSummary",
        "x-plugin-required": "gpu",
        "responses": {
          "200": {
            "description": "GPU strategy and inventory summary",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GPUSummaryResponse" }
              }
            }
          },
          "403": { "description": "User is not authorized (RBAC)" },
          "404": { "description": "Plugin disabled — enable the `gpu` plugin via ROS_ENABLED_PLUGINS" },
          "503": { "description": "Service unavailable" }
        }
      }
    },
...
    "/recommendations/openshift/gpu/mig": {
      "get": {
        "tags": ["GPU Recommendations"],
        "summary": "List MIG GPU profile recommendations",
        "description": "Returns container-level rows where the recommended GPU profile is a MIG slice (not `full_gpu`). Useful for focusing on MIG right-sizing without filtering the full container recommendation list.",
        "operationId": "getGpuMigRecommendations",
        "x-plugin-required": "gpu",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 100, "minimum": 1, "maximum": 1000 }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 0, "minimum": 0 }
          },
          {
            "name": "order_by",
            "in": "query",
            "description": "Sort field (applied in memory)",
            "required": false,
            "schema": { "type": "string", "default": "cluster_uuid", "enum": ["cluster_uuid", "namespace", "workload", "container", "term", "gpu_model", "confidence"] }
          },
          {
            "name": "order_how",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "default": "desc", "enum": ["asc", "desc"] }
          }
        ],
        "responses": {
          "200": {
            "description": "MIG-focused GPU recommendations",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GPUMIGListResponse" }
              }
            }
          },
          "400": { "description": "Bad request" },
          "403": { "description": "User is not authorized (RBAC)" },
          "404": { "description": "Plugin disabled — enable the `gpu` plugin via ROS_ENABLED_PLUGINS" },
          "503": { "description": "Service unavailable" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TermSettingsResponse": {
        "type": "object",
        "properties": {
          "recommendation_type": {
            "type": "string",
            "description": "The plugin these terms apply to"
          },
          "terms": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TermSettingsResponseItem"
            }
          }
        }
      },
      "TermSettingsResponseItem": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "enum": ["short", "medium", "long"]
          },
          "window_days": {
            "type": "integer",
            "minimum": 1,
            "description": "Maximum varies by plugin: 90 for container/namespace/node/gpu, 365 for pvc"
          },
          "min_data_days": {
            "type": "integer",
            "description": "Minimum days of data required. Auto-computed as half of window_days if not explicitly set."
          },
          "decay_halflife_hours": {
            "type": "number",
            "description": "Exponential decay half-life in hours for weighted analysis. 0 means no decay (equal weighting)."
          },
          "is_default": {
            "type": "boolean",
            "description": "True if this term uses the plugin default (no tenant or admin override)"
          },
          "locked": {
            "type": "boolean",
            "description": "True if this term is locked by administrator environment variable and cannot be modified via PUT"
          }
        }
      },
      "TermSettingsRequest": {
        "type": "object",
        "required": ["terms"],
        "properties": {
          "terms": {
            "type": "array",
            "minItems": 1,
            "maxItems": 3,
            "items": {
              "$ref": "#/components/schemas/TermSettingsRequestItem"
            }
          }
        }
      },
      "TermSettingsRequestItem": {
        "type": "object",
        "required": ["name", "window_days"],
        "properties": {
          "name": {
            "type": "string",
            "enum": ["short", "medium", "long"]
          },
          "window_days": {
            "type": "integer",
            "minimum": 1,
            "description": "Maximum varies by plugin: 90 for container/namespace/node/gpu, 365 for pvc"
          },
          "min_data_days": {
            "type": "integer",
            "minimum": 1,
            "description": "Optional. If omitted, auto-computed as half of window_days (minimum 1). Must be <= window_days."
          },
          "decay_halflife_hours": {
            "type": "number",
            "minimum": 0,
            "maximum": 8760,
            "description": "Optional. Exponential decay half-life in hours. 0 means no decay."
          }
        }
      },
...
      "GPUSummaryResponse": {
        "type": "object",
        "properties": {
          "mig": { "$ref": "#/components/schemas/GPUStrategySummary" },
          "timeslicing": { "$ref": "#/components/schemas/GPUStrategySummary" },
          "total_gpus_analyzed": {
            "type": "integer",
            "description": "Distinct (cluster, node, GPU model) triples in gpu_container_digests for visible clusters"
          },
          "clusters_with_gpu_data": {
            "type": "integer",
            "description": "Distinct clusters with at least one gpu_container_digests row"
          }
        }
      },
      "GPUStrategySummary": {
        "type": "object",
        "properties": {
          "count": { "type": "integer" },
          "link": { "type": "string", "description": "Relative URL path to the strategy listing endpoint" }
        }
      },
      "GPUMIGListResponse": {
        "type": "object",
        "properties": {
          "meta": { "$ref": "#/components/schemas/GPUMIGListMeta" },
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/GPUMIGRecommendationEntry" }
          },
          "warnings": {
            "type": "array",
            "items": { "type": "string" }
          }
        }
      },
      "GPUMIGListMeta": {
        "type": "object",
        "properties": {
          "count": { "type": "integer" },
          "limit": { "type": "integer" },
          "offset": { "type": "integer" }
        }
      },
      "GPUMIGRecommendationEntry": {
        "type": "object",
        "properties": {
          "cluster_uuid": { "type": "string" },
          "namespace": { "type": "string" },
          "workload": { "type": "string" },
          "container": { "type": "string" },
          "term": { "type": "string" },
          "gpu_model": { "type": "string" },
          "node_name": { "type": "string" },
          "recommended_gpu_profile": { "type": "string" },
          "current_gpu_profile": { "type": "string" },
          "gpu_classification": { "type": "string" },
          "confidence": { "type": "number", "format": "float" }
        }
      },
...
    }
  }
}